Refactoring and setup client for ui changes
