+++ client/package.json
@@ -1,0 +1,5 @@
+{
+  "name": "client",
+  "version": "0.1.0",
+  "private": true
+}
