+++ server/src/main/java/org/romanconversion/controller/RomanConversion.java
@@ -1,6 +1,6 @@
-package org.adobemarketing.adobeperformance.controller;
+package org.romanconversion.controller;
 
-import org.adobemarketing.adobeperformance.RomanService;
+import org.romanconversion.server.RomanService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
