+++ server/src/test/java/org/romanconversion/ServerApplicationTests.java
@@ -1,0 +1,13 @@
+package org.romanconversion;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ServerApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
