+++ server/src/test/java/org/romanconversion/controller/RomanConversionTest.java
@@ -1,8 +1,9 @@
-package org.adobemarketing.adobeperformance.controller;
+package org.romanconversion.controller;
 
-import org.adobemarketing.adobeperformance.RomanService;
+import org.romanconversion.server.RomanService;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
+import org.romanconversion.controller.RomanConversion;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.test.context.bean.override.mockito.MockitoBean;
